Fixed selldrops command, fixed placeholders, fixed upgrade gui

+++ src/main/java/xyz/arcadiadevs/infiniteforge/events/BlockPlace.java
@@ -79,7 +79,7 @@ public class BlockPlace implements Listener {
     final Player player = event.getPlayer();
 
     int tier = Integer.parseInt(firstLine.split(" ")[2]);
-    final int limit = PlayerUtil.getGeneratorLimit(player);
+    final double limit = PlayerUtil.getGeneratorLimit(player);
     final boolean enabled = config.getBoolean("limit-settings.enabled");
 
     if (locationsData.getGeneratorsCountByPlayer(player) >= limit
